refactor internal and add chimera
