+++ chimera/runtime.go
@@ -1,0 +1,6 @@
+package chimera
+
+import "github.com/flier/gohs/internal/ch"
+
+// A Chimera scratch space.
+type Scratch = ch.Scratch
