+++ hyperscan/api.go
@@ -3,44 +3,9 @@ package hyperscan
 import (
 	"fmt"
 	"io"
+
+	"github.com/flier/gohs/internal/hs"
 )
-
-type matchEvent struct {
-	id       uint
-	from, to uint64
-	flags    ScanFlag
-}
-
-func (e *matchEvent) Id() uint { return e.id } // nolint: golint,revive,stylecheck
-
-func (e *matchEvent) From() uint64 { return e.from }
-
-func (e *matchEvent) To() uint64 { return e.to }
-
-func (e *matchEvent) Flags() ScanFlag { return e.flags }
-
-type matchRecorder struct {
-	matched []matchEvent
-	err     error
-}
-
-func (h *matchRecorder) Matched() bool { return h.matched != nil }
-
-func (h *matchRecorder) Handle(id uint, from, to uint64, flags uint, context interface{}) error {
-	if len(h.matched) > 0 {
-		tail := &h.matched[len(h.matched)-1]
-
-		if tail.id == id && tail.from == from && tail.flags == ScanFlag(flags) && tail.to < to {
-			tail.to = to
-
-			return h.err
-		}
-	}
-
-	h.matched = append(h.matched, matchEvent{id, from, to, ScanFlag(flags)})
-
-	return h.err
-}
 
 // Match reports whether the byte slice b contains any match of the regular expression pattern.
 func Match(pattern string, data []byte) (bool, error) {
@@ -66,13 +31,13 @@ func Match(pattern string, data []byte) (bool, error) {
 		_ = s.Free()
 	}()
 
-	h := &matchRecorder{}
+	h := &hs.MatchRecorder{}
 
 	if err = db.Scan(data, s, h.Handle, nil); err != nil {
 		return false, err // nolint: wrapcheck
 	}
 
-	return h.Matched(), h.err
+	return h.Matched(), h.Err
 }
 
 // MatchReader reports whether the text returned by the Reader contains any match of the regular expression pattern.
@@ -99,13 +64,13 @@ func MatchReader(pattern string, reader io.Reader) (bool, error) {
 		_ = s.Free()
 	}()
 
-	h := &matchRecorder{}
+	h := &hs.MatchRecorder{}
 
 	if err = db.Scan(reader, s, h.Handle, nil); err != nil {
 		return false, err // nolint: wrapcheck
 	}
 
-	return h.Matched(), h.err
+	return h.Matched(), h.Err
 }
 
 // MatchString reports whether the string s contains any match of the regular expression pattern.
