+++ internal/ch/error.go
@@ -1,50 +1,41 @@
-package chimera
+package ch
 
-/*
-#cgo pkg-config: libch libhs
-#cgo linux LDFLAGS: -lm -lpcre
-#cgo darwin LDFLAGS: -lpcre
-
-#include <stdlib.h>
-#include <limits.h>
-#include <stdint.h>
-
-#include <ch.h>
-*/
+// #include <ch.h>
 import "C"
+
 import "fmt"
 
-// ChError represents an error
-type ChError C.ch_error_t
+// Error represents an error.
+type Error C.ch_error_t
 
 const (
 	// ErrSuccess is the error returned if the engine completed normally.
-	ErrSuccess ChError = C.CH_SUCCESS
+	ErrSuccess Error = C.CH_SUCCESS
 	// ErrInvalid is the error returned if a parameter passed to this function was invalid.
-	ErrInvalid ChError = C.CH_INVALID
+	ErrInvalid Error = C.CH_INVALID
 	// ErrNoMemory is the error returned if a memory allocation failed.
-	ErrNoMemory ChError = C.CH_NOMEM
+	ErrNoMemory Error = C.CH_NOMEM
 	// ErrScanTerminated is the error returned if the engine was terminated by callback.
-	ErrScanTerminated ChError = C.CH_SCAN_TERMINATED
+	ErrScanTerminated Error = C.CH_SCAN_TERMINATED
 	// ErrCompileError is the error returned if the pattern compiler failed.
-	ErrCompileError ChError = C.CH_COMPILER_ERROR
+	ErrCompileError Error = C.CH_COMPILER_ERROR
 	// ErrDatabaseVersionError is the error returned if the given database was built for a different version of Hyperscan.
-	ErrDatabaseVersionError ChError = C.CH_DB_VERSION_ERROR
+	ErrDatabaseVersionError Error = C.CH_DB_VERSION_ERROR
 	// ErrDatabasePlatformError is the error returned if the given database was built for a different platform (i.e., CPU type).
-	ErrDatabasePlatformError ChError = C.CH_DB_PLATFORM_ERROR
+	ErrDatabasePlatformError Error = C.CH_DB_PLATFORM_ERROR
 	// ErrDatabaseModeError is the error returned if the given database was built for a different mode of operation.
-	ErrDatabaseModeError ChError = C.CH_DB_MODE_ERROR
+	ErrDatabaseModeError Error = C.CH_DB_MODE_ERROR
 	// ErrBadAlign is the error returned if a parameter passed to this function was not correctly aligned.
-	ErrBadAlign ChError = C.CH_BAD_ALIGN
+	ErrBadAlign Error = C.CH_BAD_ALIGN
 	// ErrBadAlloc is the error returned if the memory allocator did not correctly return memory suitably aligned.
-	ErrBadAlloc ChError = C.CH_BAD_ALLOC
+	ErrBadAlloc Error = C.CH_BAD_ALLOC
 	// ErrScratchInUse is the error returned if the scratch region was already in use.
-	ErrScratchInUse ChError = C.CH_SCRATCH_IN_USE
+	ErrScratchInUse Error = C.CH_SCRATCH_IN_USE
 	// ErrUnknown is the unexpected internal error from Hyperscan.
-	ErrUnknownHSError ChError = C.CH_UNKNOWN_HS_ERROR
+	ErrUnknownHSError Error = C.CH_UNKNOWN_HS_ERROR
 )
 
-var chErrorMessages = map[ChError]string{
+var ErrorMessages = map[Error]string{
 	C.CH_SUCCESS:           "The engine completed normally.",
 	C.CH_INVALID:           "A parameter passed to this function was invalid.",
 	C.CH_NOMEM:             "A memory allocation failed.",
@@ -59,19 +50,10 @@ var chErrorMessages = map[ChError]string{
 	C.CH_UNKNOWN_HS_ERROR:  "Unexpected internal error from Hyperscan.",
 }
 
-func (e ChError) Error() string {
-	if msg, exists := chErrorMessages[e]; exists {
+func (e Error) Error() string {
+	if msg, exists := ErrorMessages[e]; exists {
 		return msg
 	}
 
-	return fmt.Sprintf("unexpected error, %d", int(e))
-}
-
-type (
-	chDatabase *C.ch_database_t
-	chScratch  *C.ch_scratch_t
-)
-
-func chVersion() string {
-	return C.GoString(C.ch_version())
+	return fmt.Sprintf("unexpected error, %d", int(C.ch_error_t(e)))
 }
