+++ internal/ch/link.go
@@ -1,0 +1,8 @@
+package ch
+
+/*
+#cgo pkg-config: libch libhs
+#cgo linux LDFLAGS: -lm -lpcre
+#cgo darwin LDFLAGS: -lpcre
+*/
+import "C"
