+++ internal/hs/link.go
@@ -1,0 +1,7 @@
+package hs
+
+/*
+#cgo pkg-config: libhs
+#cgo linux LDFLAGS: -lm
+*/
+import "C"
